Add more architectures
We have currently converted these with SIMDE:
- aarch64 (sse2, sse41, avx)
- armv7   (sse2, sse41)
- ppc64le (sse2, sse41, avx)
- sparc64 (sse2, sse41)

+++ main.c
@@ -128,7 +128,7 @@ int main(int argc, char **argv) {
     for (size_t i = 0; i < out_len; i++) {
       printf("%02x", buf[i]);
     }
-    printf(" cycles=%llu\n", stop - start);
+    printf(" cycles=%llu\n", (long long unsigned)(stop - start));
   }
   printf("  -\n");
